backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx>=0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
google-genai==1.52.0
python-multipart==0.0.18
pillow==11.0.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
